// AngularPocs-master/ADMTC-ACAD/src/app/models/user_corrector.model.ts
export class CorrectorUser {
    id: string;
    title: string;
    preparationCenter: string;
    firstname: string;
    lastname: string;
    email: string;
    telephone: string;
    userType: string;
    acadmicTitle: string;
    resume: any;
    correctorType: string;

    constructor() {
        this.id = "";
        this.title = "";
        this.preparationCenter = "";
        this.firstname = "";
        this.lastname = "";
        this.email = "";
        this.telephone = "";
        this.userType = "";
        this.acadmicTitle = "";
        this.resume = "";
        this.correctorType = "";
    }
}