// AngularPocs-master/ADMTC-ACAD/src/app/models/userTypes.model.ts
export class UserTypesModel {

    _id: string;
    name: string;
    // New Things
    // _id: string;    // autogenerated ID from database
    // name: string;   // type name
    entity: string; // 'admtc', 'academic', 'company', 'service-provider'
    permissions: {
        academicKit: {
            create: Boolean,
            read: Boolean,
            update: Boolean,
            delete: Boolean,
            test: {
                create: Boolean,
                read: Boolean,
                update: Boolean,
                delete: Boolean,
                documents: {
                    create: Boolean,
                    read: Boolean,
                    update: Boolean,
                    delete: Boolean
                }
            },
            documents: {
                create: Boolean,
                read: Boolean,
                update: Boolean,
                delete: Boolean
            }
        },
        users: {
            create: Boolean,
            read: Boolean,
            update: Boolean,
            delete: Boolean
        },
        rncpTitles: {
            create: Boolean,
            read: Boolean, // default true
            update: Boolean,
            delete: Boolean
        }
    }


    constructor() {
        this._id = '';
        this.name = '';
    }
}